backend/app/services/video_generator.py
"""
Video Generator Service
Generates videos with brand DNA and mandatory logo watermark
"""

from typing import Dict, Optional, List
from PIL import Image
import io


class VideoGeneratorService:
    """
    Service for generating brand-consistent videos with logo watermark
    
    Pipeline:
    1. Generate keyframes with brand DNA (via Image Generator)
    2. Animate with Runway Gen-3 / Luma Dream Machine
    3. Add audio track (music + optional voice-over)
    4. **ADD LOGO WATERMARK** (MANDATORY - always added)
    5. Render final video with MoviePy/FFmpeg
    """
    
    def __init__(self):
        self.runway_client = None  # Initialize with API token
        self.elevenlabs_client = None  # For voice-over
    
    async def generate_video(
        self,
        prompt: str,
        brand_dna: Dict,
        platform_specs: Dict,
        duration_seconds: int = 15,
        include_voiceover: bool = False,
        voiceover_text: Optional[str] = None
    ) -> bytes:
        """
        Generate a complete branded video with logo watermark
        """
        # Step 1: Generate keyframes
        keyframes = await self._generate_keyframes(prompt, brand_dna)
        
        # Step 2: Animate keyframes
        video_clip = await self._animate_keyframes(keyframes, duration_seconds)
        
        # Step 3: Add audio
        if include_voiceover and voiceover_text:
            video_clip = await self._add_voiceover(video_clip, voiceover_text, brand_dna)
        else:
            video_clip = await self._add_background_music(video_clip, brand_dna)
        
        # Step 4: ADD LOGO WATERMARK (MANDATORY)
        video_clip = await self._add_logo_watermark(video_clip, brand_dna)
        
        # Step 5: Render to specs
        final_video = await self._render_video(video_clip, platform_specs)
        
        return final_video
    
    async def _generate_keyframes(
        self,
        prompt: str,
        brand_dna: Dict,
        num_keyframes: int = 3
    ) -> List[Image.Image]:
        """
        Generate keyframe images for video animation
        """
        # TODO: Use ImageGeneratorService to create keyframes
        # For now, create placeholders
        
        keyframes = []
        bg_color = brand_dna.get("vector_cromatico", {}).get("primary", "#1A365D")
        
        for i in range(num_keyframes):
            img = Image.new("RGB", (1920, 1080), bg_color)
            keyframes.append(img)
        
        return keyframes
    
    async def _animate_keyframes(
        self,
        keyframes: List[Image.Image],
        duration_seconds: int
    ) -> Dict:
        """
        Animate keyframes using Runway Gen-3 or Luma Dream Machine
        
        Returns a video clip object for further processing.
        """
        # TODO: Implement with Runway API
        # The API call would look something like:
        #
        # async with httpx.AsyncClient() as client:
        #     response = await client.post(
        #         "https://api.runwayml.com/v1/generate",
        #         headers={"Authorization": f"Bearer {RUNWAY_API_KEY}"},
        #         json={
        #             "model": "gen-3-alpha",
        #             "image": base64_keyframe,
        #             "motion": "smooth_zoom_in",
        #             "duration": duration_seconds
        #         }
        #     )
        
        return {
            "frames": keyframes,
            "duration": duration_seconds,
            "fps": 30,
            "audio": None
        }
    
    async def _add_voiceover(
        self,
        video_clip: Dict,
        text: str,
        brand_dna: Dict
    ) -> Dict:
        """
        Add AI voice-over using ElevenLabs
        
        The voice style should match the brand's tone.
        """
        # TODO: Implement with ElevenLabs API
        # Select voice based on brand tone (formal/casual, etc.)
        
        tone = brand_dna.get("vector_semantico", {}).get("emocion", "profesional")
        
        # voice_mapping = {
        #     "profesional": "voice_id_professional",
        #     "amigable": "voice_id_friendly",
        #     "autoritativo": "voice_id_authoritative"
        # }
        
        video_clip["audio"] = {
            "type": "voiceover",
            "text": text,
            "tone": tone
        }
        
        return video_clip
    
    async def _add_background_music(
        self,
        video_clip: Dict,
        brand_dna: Dict
    ) -> Dict:
        """
        Add background music matching brand mood
        """
        # TODO: Select music from stock library based on brand tone
        
        video_clip["audio"] = {
            "type": "music",
            "mood": brand_dna.get("vector_semantico", {}).get("emocion", "profesional")
        }
        
        return video_clip
    
    async def _add_logo_watermark(
        self,
        video_clip: Dict,
        brand_dna: Dict
    ) -> Dict:
        """
        ADD LOGO WATERMARK TO VIDEO - MANDATORY STEP
        
        This method MUST be called for every generated video.
        The logo appears as a watermark throughout the video.
        """
        logo_config = brand_dna.get("logo", {})
        
        if not logo_config.get("logo_url"):
            print("WARNING: No logo configured for brand. Videos should have logo watermark!")
            return video_clip
        
        # Configure watermark settings
        watermark_config = {
            "logo_url": logo_config.get("logo_url"),
            "position": logo_config.get("preferred_position", "bottom-right"),
            "opacity": 0.85,  # Slightly transparent for watermark effect
            "size_percent": logo_config.get("max_size_percent", 0.12),
            "padding_percent": logo_config.get("padding_percent", 0.03),
            "fade_in": True,  # Fade in at start
            "fade_in_duration": 0.5,  # seconds
        }
        
        video_clip["watermark"] = watermark_config
        video_clip["has_logo"] = True  # ALWAYS true after this step
        
        return video_clip
    
    async def _render_video(
        self,
        video_clip: Dict,
        platform_specs: Dict
    ) -> bytes:
        """
        Render final video with MoviePy/FFmpeg
        """
        # TODO: Implement actual rendering with MoviePy
        # 
        # from moviepy.editor import (
        #     ImageSequenceClip, CompositeVideoClip, 
        #     ImageClip, AudioFileClip
        # )
        #
        # # Create base video from frames
        # base_clip = ImageSequenceClip(frames, fps=30)
        #
        # # Add logo watermark
        # if video_clip.get("watermark"):
        #     logo = ImageClip(logo_path).set_position(position)
        #     logo = logo.set_opacity(opacity)
        #     base_clip = CompositeVideoClip([base_clip, logo])
        #
        # # Add audio
        # if video_clip.get("audio"):
        #     audio = AudioFileClip(audio_path)
        #     base_clip = base_clip.set_audio(audio)
        #
        # # Resize to specs
        # base_clip = base_clip.resize((width, height))
        #
        # # Render
        # base_clip.write_videofile(output_path, codec="libx264")
        
        # Placeholder return
        return b"placeholder_video_bytes"
    
    async def generate_animated_image(
        self,
        image: Image.Image,
        brand_dna: Dict,
        animation_type: str = "subtle_zoom"
    ) -> bytes:
        """
        Create a simple animation from a static image
        
        Useful for turning product photos into engaging video content.
        Animation types: subtle_zoom, pan, parallax
        """
        # TODO: Implement with Runway Image-to-Video
        
        # Even animated images get the logo watermark
        return await self.generate_video(
            prompt="Animate this image",
            brand_dna=brand_dna,
            platform_specs={"width": 1080, "height": 1920, "format": "mp4"},
            duration_seconds=5
        )
