backend/requirements.txt
# FastAPI Backend - Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.25

# AI/ML
openai==1.12.0
langchain==0.1.5
langgraph==0.0.20
replicate==0.23.0

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3

# Image Processing
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Video Processing
moviepy==1.0.3

# Vector Database
pinecone-client==3.0.0
